# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: OpenRating API
  version: 1.0.0
servers:
  - url: https://api.example.com/v1
paths:
...
  /v1/events:
    post:
      tags: [Events]
      summary: Create an event
      description: Create a tournament, league, or other competition container for matches.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreate'
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                type: object
                properties:
                  event:
                    $ref: '#/components/schemas/Event'
        '400':
          description: Invalid request
    get:
      tags: [Events]
      summary: List events for an organization
      parameters:
        - in: query
          name: organization_id
          required: false
          schema: { type: string }
          description: UUID of the organization (optional if organization_slug provided)
        - in: query
          name: organization_slug
          required: false
          schema: { type: string }
          description: Slug identifier for the organization
        - in: query
          name: types
          required: false
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
              enum: [TOURNAMENT, LEAGUE, LADDER, BOX_LEAGUE, CHAMPIONSHIP, SERIES, EXHIBITION, CUSTOM]
        - in: query
          name: q
          required: false
          schema: { type: string }
          description: Case-insensitive search over event name or slug
        - in: query
          name: cursor
          required: false
          schema: { type: string }
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        '200':
          description: Paginated list of events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
  /v1/events/{event_id}:
    get:
      tags: [Events]
      summary: Retrieve an event by ID
      parameters:
        - in: path
          name: event_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                type: object
                properties:
                  event:
                    $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
    patch:
      tags: [Events]
      summary: Update an event
      parameters:
        - in: path
          name: event_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventUpdate'
      responses:
        '200':
          description: Updated event
          content:
            application/json:
              schema:
                type: object
                properties:
                  event:
                    $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
  /v1/events/{event_id}/participants:
    post:
      tags: [Events]
      summary: Upsert an event participant
      parameters:
        - in: path
          name: event_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventParticipantRequest'
      responses:
        '201':
          description: Participant ensured
          content:
            application/json:
              schema:
                type: object
                properties:
                  participant:
                    $ref: '#/components/schemas/EventParticipant'
        '404':
          description: Event not found
    get:
      tags: [Events]
      summary: List participants for an event
      parameters:
        - in: path
          name: event_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Event participants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventParticipantListResponse'
        '404':
          description: Event not found
...
components:
  schemas:
...
    Event:
      type: object
      properties:
        event_id: { type: string }
        organization_id: { type: string }
        organization_slug: { type: string }
        type:
          type: string
          enum: [TOURNAMENT, LEAGUE, LADDER, BOX_LEAGUE, CHAMPIONSHIP, SERIES, EXHIBITION, CUSTOM]
        name: { type: string }
        slug: { type: string }
        description:
          type: [string, 'null']
        start_date:
          type: [string, 'null']
          format: date-time
        end_date:
          type: [string, 'null']
          format: date-time
        classification:
          $ref: '#/components/schemas/EventClassification'
        sanctioning_body:
          type: [string, 'null']
        season:
          type: [string, 'null']
        purse:
          type: [number, 'null']
        purse_currency:
          type: [string, 'null']
          description: ISO-4217 code representing the purse currency.
        media_links:
          $ref: '#/components/schemas/EventMediaLinks'
        metadata:
          type: [object, 'null']
          additionalProperties: true
        competitions:
          type: array
          description: Nested competitions that belong to this event.
          items:
            $ref: '#/components/schemas/Competition'
        created_at:
          type: [string, 'null']
          format: date-time
        updated_at:
          type: [string, 'null']
          format: date-time
    EventCreate:
      type: object
      required: [type, name]
      properties:
        organization_id: { type: string, format: uuid }
        organization_slug: { type: string }
        type:
          type: string
          enum: [TOURNAMENT, LEAGUE, LADDER, BOX_LEAGUE, CHAMPIONSHIP, SERIES, EXHIBITION, CUSTOM]
        name: { type: string }
        slug: { type: string }
        description:
          type: [string, 'null']
        start_date:
          type: [string, 'null']
          format: date-time
        end_date:
          type: [string, 'null']
          format: date-time
        classification:
          $ref: '#/components/schemas/EventClassification'
        sanctioning_body:
          type: [string, 'null']
        season:
          type: [string, 'null']
        purse:
          type: [number, 'null']
        purse_currency:
          type: [string, 'null']
        media_links:
          $ref: '#/components/schemas/EventMediaLinks'
        metadata:
          type: [object, 'null']
          additionalProperties: true
    EventUpdate:
      type: object
      properties:
        type:
          type: string
          enum: [TOURNAMENT, LEAGUE, LADDER, BOX_LEAGUE, CHAMPIONSHIP, SERIES, EXHIBITION, CUSTOM]
        name: { type: string }
        slug: { type: string }
        description:
          type: [string, 'null']
        start_date:
          type: [string, 'null']
          format: date-time
        end_date:
          type: [string, 'null']
          format: date-time
        classification:
          $ref: '#/components/schemas/EventClassification'
        sanctioning_body:
          type: [string, 'null']
        season:
          type: [string, 'null']
        purse:
          type: [number, 'null']
        purse_currency:
          type: [string, 'null']
        media_links:
          $ref: '#/components/schemas/EventMediaLinks'
        metadata:
          type: [object, 'null']
          additionalProperties: true
    Competition:
      type: object
      properties:
        competition_id: { type: string }
        event_id: { type: string }
        event_slug: { type: string }
        organization_id: { type: string }
        name: { type: string }
        slug: { type: string }
        sport:
          type: [string, 'null']
          enum: [BADMINTON, TENNIS, SQUASH, PADEL, PICKLEBALL]
        discipline:
          type: [string, 'null']
          enum: [SINGLES, DOUBLES, MIXED]
        format:
          type: [string, 'null']
        tier:
          type: [string, 'null']
        status:
          type: [string, 'null']
        draw_size:
          type: [integer, 'null']
        start_date:
          type: [string, 'null']
          format: date-time
        end_date:
          type: [string, 'null']
          format: date-time
        metadata:
          type: [object, 'null']
          additionalProperties: true
        created_at:
          type: [string, 'null']
          format: date-time
        updated_at:
          type: [string, 'null']
          format: date-time
...
    EventClassification:
      type: object
      description: Structured hierarchy data describing the competitive level of an event.
      properties:
        level:
          $ref: '#/components/schemas/EventLevel'
          description: High-level tour or competition grade (e.g. WORLD_TOUR, NATIONAL, CLUB).
        grade:
          $ref: '#/components/schemas/EventGrade'
          description: Fine-grained grade within the level (e.g. SUPER_750, MAJOR, DIVISION_1).
        age_group:
          $ref: '#/components/schemas/EventAgeGroup'
          description: Target age bracket if applicable (e.g. U19, SENIOR).
        tour:
          type: [string, 'null']
          description: Named series or tour identifier.
        category:
          type: [string, 'null']
          description: Discipline-specific category, draw, or bracket label.
    EventMediaLinks:
      type: object
      description: Links to event resources across digital channels.
      properties:
        website:
          type: [string, 'null']
          format: uri
        registration:
          type: [string, 'null']
          format: uri
        live_scoring:
          type: [string, 'null']
          format: uri
        streaming:
          type: [string, 'null']
          format: uri
        social:
          type: [object, 'null']
          additionalProperties: { type: string, format: uri }
    EventListResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        next_cursor:
          type: [string, 'null']
    EventParticipant:
      type: object
      properties:
        event_id: { type: string }
        player_id: { type: string }
        seed:
          type: [integer, 'null']
        status:
          type: [string, 'null']
        metadata:
          type: [object, 'null']
          additionalProperties: true
        created_at:
          type: [string, 'null']
          format: date-time
        updated_at:
          type: [string, 'null']
          format: date-time
    EventParticipantRequest:
      type: object
      required: [player_id]
      properties:
        player_id: { type: string }
        seed:
          type: [integer, 'null']
        status:
          type: [string, 'null']
        metadata:
          type: [object, 'null']
          additionalProperties: true
    EventParticipantListResponse:
      type: object
      properties:
        event:
          $ref: '#/components/schemas/Event'
        participants:
          type: array
          items:
            $ref: '#/components/schemas/EventParticipant'
...
    EventLevel:
      type: [string, 'null']
      description: Top-level classification tier for an event.
      enum: [WORLD_TOUR, CONTINENTAL, NATIONAL, REGIONAL, CLUB, SCHOOL, COMMUNITY, OTHER, null]
    EventGrade:
      type: [string, 'null']
      description: Grade within a level, often tied to ranking points.
      enum: [SUPER_1000, SUPER_750, SUPER_500, SUPER_300, GOLD, SILVER, BRONZE, MAJOR, DIVISION_1, DIVISION_2, OPEN, OTHER, null]
    EventAgeGroup:
      type: [string, 'null']
      description: Target age bracket for participants.
      enum: [U11, U13, U15, U17, U19, U21, SENIOR, ADULT, VETERAN, MASTER, OPEN, OTHER, null]
...